Extract WorkPackageElement to its own file
