+++ lib/services/css_constants.ts
@@ -1,0 +1,4 @@
+export const SPACER_S = "4px";
+export const SPACER_M = "8px";
+export const SPACER_L = "12px";
+export const SPACER_XL = "16px";
